# Avoid most allocations in GetExtensionFromUri