@PhyxionNL 2021-11-12
Added a second commit to remove most allocations from this method if running .NET Core 3.1 and higher. Maybe drop .NET Core 2.1 target? It's end of life.