@JimBobSquarePants 2021-11-17
This was fun! Thanks both of you!